refactor(aggregator): move `test::mocks` to `test::double::mocks`
As mocks are test double, it should have been there in the first place
