+++ internal/cardano-node/mithril-cardano-node-chain/src/test/mod.rs
@@ -8,4 +8,5 @@ pub mod double;
 #[cfg(test)]
 pub(crate) mod test_cli_runner;
 
+#[cfg(test)]
 mithril_common::define_test_logger!();
