+++ internal/mithril-dmq/src/test/mod.rs
@@ -9,4 +9,5 @@ pub mod double;
 #[cfg(test)]
 pub(crate) mod payload;
 
+#[cfg(test)]
 mithril_common::define_test_logger!();
