+++ mithril-aggregator/src/test/double/mod.rs
@@ -1,1 +1,7 @@
+//! Test doubles
+//!
+//! Enable unit testing with controlled inputs and predictable behavior.
+
 mod dummies;
+#[cfg(test)]
+pub mod mocks;
